the pointer address must be cast to an unsigned long to operate properly... unfortunately this went unnoticed during my mutexpool enhancements :/
[git-p4: depot-paths = "//depot/qt/main/": change = 78317]

+++ src/tools/qmutexpool.cpp
@@ -133,7 +133,7 @@ QMutexPool::~QMutexPool()
 */
 QMutex *QMutexPool::get( void *address )
 {
-    int index = (int) address % count;
+    int index = (int) ( (unsigned long) address % count );
 
     if ( ! mutexes[index] ) {
 	// mutex not created, create one
